Fix iterator type mismatch

+++ vcu/App/Drivers/Src/pdu.cpp
@@ -23,7 +23,7 @@ void PDU::setCurrentLimit(uint8_t channel, float amps) {
 	if (channel != 0 && channel <= NUM_CHANNELS) { // check channel in range
 		PDU_MAX_CURRENT = CHANNEL_MASK & (1 << (channel - 1)) ? LOW_CURRENT_LIMIT : HIGH_CURRENT_LIMIT; // set to 10A if 1, 20 if 0
 		state.requestedCurrentLimit[channel - 1] = std::min(amps, PDU_MAX_CURRENT);
-		for (int i = 0; i < NUM_CHANNELS; i++) {
+		for (size_t i = 0; i < NUM_CHANNELS; i++) {
 			txData[i] = static_cast<uint8_t>(
 				std::round(state.requestedCurrentLimit[i] * PDU_BIT_TO_POWER_SCALE)
 				);
@@ -33,7 +33,7 @@ void PDU::setCurrentLimit(uint8_t channel, float amps) {
 void PDU::setPWMDutyCycle(uint8_t channel, uint8_t dutyCyclePercent) {
 	if (channel != 0 && channel <= NUM_CHANNELS) {
 		state.requestedPWMDutyPercent[channel - 1] = std::min(dutyCyclePercent, PDU_MAX_PWM);
-		for (int i = 0; i < NUM_CHANNELS; i++) {
+		for (size_t i = 0; i < NUM_CHANNELS; i++) {
 			txData[i] = static_cast<uint8_t>(
 				std::round(state.requestedPWMDutyPercent[i] * PDU_BIT_TO_POWER_SCALE)
 				);
@@ -60,7 +60,7 @@ void PDU::processMessage(int channelStart, const can::Message &message) {
 	if (message.numBytes != 8) {
 		return; // incorrect number of bytes received, bad message
 	}
-	for (int i = 0; i < NUM_CHANNELS; i += 2) {
+	for (size_t i = 0; i < NUM_CHANNELS; i += 2) {
 		uint8_t errorCode = ((message.data[i] & ERROR_MASK) >> 5);
 		switch (errorCode) {
 		case 0x00:
